server/requirements.txt
Django==1.8.3
psycopg2==2.6.1
djangorestframework==3.2